# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MockCam Management API
  version: 1.5.0
  summary: REST, WebSocket and media endpoints of the MockCam virtual network camera emulator.
  description: |
    MockCam emulates an ONVIF Profile S network camera for VMS / NVR development and load testing.
    This document describes the **management API** served on the HTTP port (default `8080`).

    * The management API does not require authentication. RTSP and the ONVIF SOAP services honour
      `server.auth_type` (Digest / Basic / none).
    * Request bodies are JSON, at most 1 MiB, and **unknown fields are rejected with `400`** so that
      typos in setting names are surfaced.
    * Errors use the `Error` schema: `{"error": "<message>"}`.
    * `405 Method Not Allowed` is returned for unsupported methods.

    Related documents: [REST API notes (Japanese)](rest_api.md), [ONVIF Profile S](onvif_profile_s.md).
  license:
    name: MIT
    identifier: MIT
  contact:
    name: MockCam
    url: https://github.com/miyabiver39/mockcam
...
servers:
  - url: http://localhost:8080
    description: Local instance (default http_port)
...
paths:
...
  /api/config:
    get:
      tags: [System]
      operationId: getConfig
      summary: Full configuration (settings.json)
      responses:
        "200":
          description: Configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
    put:
      tags: [System]
      operationId: updateServerConfig
      summary: Update the server section
      description: |
        Only the `server` section is applied; `profiles` and `ptz` in the body are ignored.
        Ports, authentication and log level are validated (`400` on error). Including `log_level`
        changes the runtime log level immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                server:
                  $ref: "#/components/schemas/ServerConfig"
              required: [server]
      responses:
        "200":
          description: Applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: ok }
                  server: { $ref: "#/components/schemas/ServerConfig" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/logs:
    get:
      tags: [System]
      operationId: getLogs
      summary: Recent log entries (last 200)
      responses:
        "200":
          description: Log entries, oldest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LogEntry" }
    put:
      tags: [System]
      operationId: setLogLevel
      summary: Change the log level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                level: { $ref: "#/components/schemas/LogLevel" }
              required: [level]
      responses:
        "200":
          description: Applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: ok }
                  log_level: { $ref: "#/components/schemas/LogLevel" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/profiles:
    get:
      tags: [Profiles]
      operationId: listProfiles
      summary: List profiles
      responses:
        "200":
          description: Profiles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Profile" }
    post:
      tags: [Profiles]
      operationId: createProfile
      summary: Create a profile
      description: Validates the profile, persists it and starts its FFmpeg worker.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Profile" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfileMutationResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: A profile with the same token exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/profiles/{token}:
    parameters:
      - $ref: "#/components/parameters/Token"
    get:
      tags: [Profiles]
      operationId: getProfile
      summary: Get a profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Profiles]
      operationId: updateProfile
      summary: Update a profile (hot reload)
      description: |
        Persists the profile, closes its RTSP stream so players reconnect with the new SDP, and
        restarts **only** this profile's FFmpeg worker (`SIGTERM`, then `Kill`). The `token` in the
        body is ignored; the path is authoritative.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Profile" }
      responses:
        "200":
          description: Updated and restarted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfileMutationResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Profiles]
      operationId: deleteProfile
      summary: Delete a profile
      description: Stops the worker and closes the stream. The last remaining profile cannot be deleted.
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: ok }
                  message: { type: string }
                  token: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/audio/timesignal:
    get:
      tags: [Media]
      operationId: getTimeSignalPCM
      summary: Speaking-clock PCM feed (internal)
      description: |
        Endless raw PCM (`audio/l16; rate=48000; channels=1`) used by workers whose audio mode is
        `time_signal_ja` / `time_signal_en`. Every 10-second block announces the upcoming mark and plays
        the tone pattern "ピ・ポ・ピ・ポ・ピ・ポ・ピ・ピ・ピ／" (880/440 Hz pips, 880 Hz mark).
      parameters:
        - name: lang
          in: query
          schema: { type: string, enum: [ja, en], default: ja }
      responses:
        "200":
          description: PCM stream
          content:
            audio/l16:
              schema: { type: string, format: binary }
  /api/ptz:
    get:
      tags: [PTZ]
      operationId: getPTZ
      summary: Current PTZ position
      responses:
        "200":
          description: Position
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PTZStatus" }
    post:
      tags: [PTZ]
      operationId: movePTZ
      summary: Move the virtual camera
      description: |
        `absolute` jumps to pan/tilt/zoom, `continuous` moves with the given velocity vector
        (10 % of the range per second at velocity 1.0) until `stop`. Values are clamped to
        pan/tilt `-1..1` and zoom `0..1`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PTZCommand" }
            examples:
              absolute:
                value: { action: absolute, pan: 0.5, tilt: -0.2, zoom: 0.4 }
              continuous:
                value: { action: continuous, vel_pan: 1.0, vel_tilt: 0, vel_zoom: 0 }
              stop:
                value: { action: stop }
      responses:
        "200":
          description: Position after the command
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PTZStatus"
                  - type: object
                    properties:
                      status: { type: string, const: ok }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/ptz/presets:
    get:
      tags: [PTZ]
      operationId: listPresets
      summary: Saved PTZ presets
      responses:
        "200":
          description: Presets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PTZPreset" }
    post:
      tags: [PTZ]
      operationId: managePresets
      summary: Save / go to / delete a preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action: { type: string, enum: [save_current, goto, delete] }
                name:
                  type: string
                  description: Preset name. Optional for `save_current` (defaults to `Preset_N`); an existing name is overwritten.
              required: [action]
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: ok }
                  presets:
                    type: array
                    items: { $ref: "#/components/schemas/PTZPreset" }
                  target: { $ref: "#/components/schemas/PTZPreset" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  parameters:
    Token:
      name: token
      in: path
      required: true
      description: Profile token (`[A-Za-z0-9_-]{1,64}`), e.g. `Profile_1`.
      schema: { type: string, pattern: "^[A-Za-z0-9_-]{1,64}$" }
  responses:
    BadRequest:
      description: Validation error or malformed JSON
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
    LogLevel:
      type: string
      enum: [DEBUG, INFO, WARN, ERROR]
    LogEntry:
      type: object
      properties:
        timestamp: { type: string, example: "2026-09-15 10:00:00.000" }
        level: { $ref: "#/components/schemas/LogLevel" }
        source: { type: string, example: rtsp }
        message: { type: string }
...
    DeviceInfo:
      type: object
      properties:
        manufacturer: { type: string }
        model: { type: string }
        firmware_version: { type: string, description: Always synced to the running MockCam version }
        serial_number: { type: string }
        hardware_id: { type: string }
    ServerConfig:
      type: object
      properties:
        rtsp_port: { type: integer, minimum: 1, maximum: 65535, default: 8554 }
        http_port: { type: integer, minimum: 1, maximum: 65535, default: 8080 }
        onvif_port: { type: integer, minimum: 1, maximum: 65535, default: 3702, description: WS-Discovery multicast port }
        auth_type: { type: string, enum: [digest, basic, none], default: digest, description: Applies to RTSP and ONVIF SOAP }
        auth_user: { type: string, default: admin }
        auth_pass: { type: string, default: admin1234 }
        log_level: { $ref: "#/components/schemas/LogLevel" }
        device_info: { $ref: "#/components/schemas/DeviceInfo" }
    Resolution:
      type: object
      properties:
        width: { type: integer, minimum: 0, maximum: 7680 }
        height: { type: integer, minimum: 0, maximum: 4320 }
    VideoConfig:
      type: object
      properties:
        codec: { type: string, enum: [H264, H265, HEVC, VP9, AV1], default: H264 }
        resolution: { $ref: "#/components/schemas/Resolution" }
        framerate: { type: integer, minimum: 0, maximum: 240, default: 30 }
        gop_size: { type: integer, minimum: 0, maximum: 1000, default: 30 }
        bitrate_mode: { type: string, enum: [CBR, VBR], default: CBR }
        bitrate_limit_kbps: { type: integer, minimum: 0, maximum: 200000, default: 4000 }
        quality: { type: number, description: Reserved (not used by the encoder yet) }
        pattern: { type: string, enum: [testsrc2, smptebars, allrgb, mptestsrc], description: Test pattern for `generate` mode }
        osd_text: { type: string, description: Custom overlay text (top-left) }
        show_clock: { type: boolean, description: Draw the ISO 8601 clock with milliseconds (always drawn when osd_text is empty) }
        enable_noise: { type: boolean, description: Add sensor grain }
        enable_motion_box: { type: boolean, description: Draw a moving red box for motion-detection tests }
    AudioConfig:
      type: object
      properties:
        enabled: { type: boolean, default: true }
        mode:
          type: string
          enum: [time_signal, time_signal_ja, time_signal_en, silent, noise, chime]
          default: time_signal
          description: |
            `time_signal` 880 Hz beep · `time_signal_ja`/`time_signal_en` spoken 117 clock with tones ·
            `silent` · `noise` white noise · `chime`
        codec: { type: string, enum: [AAC, G711A, PCMA, ALAW, G711U, PCMU, MULAW, G726, ADPCM_G726], default: AAC }
        bitrate_kbps: { type: integer, minimum: 0, maximum: 1024, default: 128 }
        sample_rate: { type: integer, minimum: 0, maximum: 192000, default: 44100 }
    Profile:
      type: object
      properties:
        token:
          type: string
          pattern: "^[A-Za-z0-9_-]{1,64}$"
          description: Identifier used in the RTSP path `/live/<token>` and as the ONVIF profile token
        name: { type: string }
        source_mode: { type: string, enum: [generate, file], default: generate }
        source_path: { type: string, description: Video file to loop when source_mode is `file` (falls back to generate when missing) }
        video: { $ref: "#/components/schemas/VideoConfig" }
        audio: { $ref: "#/components/schemas/AudioConfig" }
      required: [token]
    ProfileMutationResult:
      type: object
      properties:
        status: { type: string, const: ok }
        message: { type: string }
        profile: { $ref: "#/components/schemas/Profile" }
    PTZPreset:
      type: object
      properties:
        name: { type: string }
        pan: { type: number, minimum: -1, maximum: 1 }
        tilt: { type: number, minimum: -1, maximum: 1 }
        zoom: { type: number, minimum: 0, maximum: 1 }
    PTZConfig:
      type: object
      properties:
        enabled: { type: boolean }
        node_token: { type: string, default: PTZNode_1 }
        pan: { type: number, minimum: -1, maximum: 1 }
        tilt: { type: number, minimum: -1, maximum: 1 }
        zoom: { type: number, minimum: 0, maximum: 1 }
        speed: { type: number, description: Reserved }
        presets:
          type: array
          items: { $ref: "#/components/schemas/PTZPreset" }
    Config:
      type: object
      properties:
        server: { $ref: "#/components/schemas/ServerConfig" }
        profiles:
          type: array
          items: { $ref: "#/components/schemas/Profile" }
        ptz: { $ref: "#/components/schemas/PTZConfig" }
    PTZStatus:
      type: object
      properties:
        pan: { type: number }
        tilt: { type: number }
        zoom: { type: number }
        is_moving: { type: boolean }
    PTZCommand:
      type: object
      properties:
        action: { type: string, enum: [absolute, continuous, stop] }
        pan: { type: number }
        tilt: { type: number }
        zoom: { type: number }
        vel_pan: { type: number }
        vel_tilt: { type: number }
        vel_zoom: { type: number }
      required: [action]
...